Split the TagConfig class into partials: 6,427 -> 3,325 lines
Mark TagConfig `partial` and relocate two cohesive member clusters into
partial files (same class, shared members, identical behaviour):

- Core/TagConfig.Tag7.cs     — TAG7 rich-narrative builder, per-segment
                               styling, display presets, warning eval
                               (2,530 lines)
- Core/TagConfig.Defaults.cs — existing-tag index, SEQ sidecar
                               persistence, built-in default lookup maps
                               (618 lines)

Verified at each cut: both halves brace-balanced, the full member-
declaration set is byte-identical to the pre-split file (nothing lost or
duplicated), seams land on clean member boundaries, and partial class
TagConfig resolves across exactly the 3 files. BuildAndWriteTag stays in
the core file; WriteTag7All moves to the Tag7 partial.

Across the session TagConfig.cs dropped from 7,877 to 3,325 lines (-58%).
Needs a Windows dotnet build to confirm.
